Create PyroKatydid1v2
Newer version of my update to the map

CHANGELOG

-Removed bugs with unneeded fire doors screwing up the system of doors I intended to have in place
-Added more air alarms
-Fixed an issue where the vents I added were stuck on sucking out air because I am an idiot and chose the wrong vent
-Fixed a problem with the Mining Shuttle's APC not being connected to the network due to the orientation of the power cable

THINGS TO THINK ABOUT FOR THE FUTURE

-Blast doors on the reactor room to shield out radiation/explosions

The blast doors are easy enough to put in and wire to a button. The issue is where to put it. The control room is the ideal place, but it is too cramped. I would need to shift EVERYTHING along the vertical axis in that general area in order to widen the ship enough to accommodate the extra room. It is easy enough in theory but is a pain in the ass to do right. I would also need to shift the escape pods in that area to the sides and shift their corridors a little bit, and I am not sure what all that would break if I did so.